[Improve]課題３

+++ app/controllers/book_comments_controller.rb
@@ -8,10 +8,11 @@ def create
 end
 
 def destroy
-		book = Book.find(params[:book_id])
-    comment = current_user.book_comments.new(book_comment_params)
-    comment.book_id = book.id
-    comment.destroy
+	book = Book.find(params[:book_id])
+    comment = BookComment.find_by(book_id: book.id)
+    if comment.user_id == current_user.id
+        comment.destroy
+    end
     redirect_to book_path(book)
 end
